Pushing the External Service Composition Wizard Plugin
This is a major update. It includes all the functionality that allows
S-CASE developers to create from scratch a CIM model of a RESTClient
Service that will interoperate with an external, existing service once
the envisioned system is produced. This goes from zero up to and
including an External Service Layer CIM model.