Added the PSM to code transformation for External Services Compositions
The MDE engine now supports the generation of the needed code to
interact with external services that may require query parameters,
input of type JSON/XML with primitive data and/or output of type
JSON/XML with primitive data.

It does not yet support persistence to database and complex datatypes
as input/output models.