MDE 2nd Run logic
This is a major update that allows S-CASE developers to select to
re-rerun the „Generate Code“ functionality by taking as input the ecore
models produced in the previous MDE run. This pretty much drastically
reduces the amount of work the S-CASE developer has to do when he
simple wants to make a minor addition or modification to his previous
work.

This does not apply to changed done at requirements level yet.