Bug fixes (Collection Naming Properties, Multiple variables named as Query and problems with Integers and Doubles Set definitions)
This commit fixes the following issues:

1) The S-CASE developer cannot select as naming property a property
that is Collection. This lead previously to runtime errors that
violated the semantics of the naming property.

2) The naming of the „Query“ variables in deleteAllCollections function
of resource model source files, is fixed so as to able to have more
than one, one fore every resource property that is collection. Now the
name of the resource and the property prefixes the „Query“ word.

3) Finally, now the MDE engine correctly defines Sets of Doubles and
Integers as Set<Double> and Set<Integer> instead of Set<double> and
Set<int> that is not allowed in Java.