Major update for MDE external service composition
This update includes the following:

1) Complex data type handle by the MDE external service composition
component. Now the S-CASE developer can recursively define complex
datatypes to be used as input/output parameters types.

2) Loading external compositions produced by the CERTH’s composition
editor. If the compositions info are added in the projects ontology,
then the MDE engine loads it to the External service composition wizard
for editing (if it is needed) and the produces the code that
interoperates with it.

3) The  MDEs external service composition meta-models are also updated
to handle the complex data types.