Several bugfixes and database support extension MDE update.
This update provides the following:
- Fixed the External Composition MDE graphics in Windows OS.
- Fixed an issue that occurred when a searchable resource's name had at
least one capital letter and lead to compilation errors.
- Fixed an issue that lead to loading the wrong core CIM XMI file once
the re-generate button was clicked (2nd run logic of MDE).
- Fixed an issue that occurred when the user clicked the re-generate
button and renamed a resource or property, which were used in the
Authentication model, just by capitalizing (or the opposite) some
letters and lead MDE to crash.
- Fixed an issue that occurred when MDE unexpectedly crashed and made
the internal YML file uneditable without Eclipse restart.
- Fixed an issue that lead the MDE engine to always produce services
that used the same database. Now MDE creates a database that is named
as the provided Web Service name.
- Added support for "Long" datatypes in the External Composition MDE
Wizard.
- Added support for MySQL databases. At the moment the supported
databases are MySQL and PostgreSQL. The user has to provide the MDE
engine with a valid database user that has the rights to: CREATE a
database and perform all actions on it (INSERT, SELECT, DELETE, UPDATE)
- Added canceling support for all the MDE Wizards.