MDE Major Update - ABAC Release and Bug fixes
This is a major MDE plugin update. It provides the following:

1) ABAC functionality throughout the MDE engine (Wizard to code). It
must be noted that this is an alpha version and might include logic
errors regarding ABAC rules for authorization purposes.

2) Fix a bug that lead to many compilation errors when the S-CASE
developer designed a resource, some parent resource of which had also
multiple parent resources.

3) Fixed a bug that allowed the S-CASE developer to create CRUD
resources without Create or Read activities.

4) Fixed a UI bug that lead to improper resource property datatype
display, when the property datatype was anything else other than
String, Integer or Double.